refactor(git): extract cache readers into service_cache.go (#328)
